pyproject.toml
[project]
name = "encoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "streamlit>=1.47.1",
]
